feat(391-P1): PR C — relocate runtime-free core to src/core, server wraps + injects
src/core now holds only createAgent.ts + inputAssetIntake.ts + piChatSessionService.ts + index.ts (no fastify/server/Pi/cwd/process.env). Server createAgent.ts injects the runtime+service into core and owns adapter disposal; routes consume the core PiChatSessionService interface. /core Fastify-graph scan clean (4 dist files, zero fastify).
