+++ packages/agent/src/server/__tests__/createAgent.pure.test.ts
@@ -2,9 +2,9 @@ import { mkdtemp, readdir, rm } from 'node:fs/promises'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import type { AgentSessionEvent } from '@mariozechner/pi-coding-agent'
-import { createAgent } from '@hachej/boring-agent/core'
 import { describe, expect, it } from 'vitest'
 
+import { createAgent } from '../createAgent'
 import type { AgentHarness, AgentHarnessFactoryInput, AgentSendInput, RunContext } from '../../shared/harness'
 import type { AgentEvent } from '../../shared/events'
 import type { AgentTool } from '../../shared/tool'
