+++ packages/agent/src/server/__tests__/createAgent.test.ts
@@ -9,21 +9,23 @@ import {
   AGENT_NOT_IMPLEMENTED_UNTIL_T1,
   AgentFilesystemRequiredError,
   AgentNotImplementedError,
-  createAgent,
+  createAgent as createCoreAgent,
 } from '@hachej/boring-agent/core'
+import type { AgentCoreRuntimeFactory, PiChatEventSubscriber, PiChatSessionService, PiSessionRequestContext } from '@hachej/boring-agent/core'
+import { createAgent } from '../createAgent'
 import type { AgentHarness, AgentHarnessFactoryInput, AgentSendInput, RunContext } from '../../shared/harness'
 import type { SessionCtx, SessionDetail, SessionStore, SessionSummary } from '../../shared/session'
 import type { PiAgentPromptInput, PiAgentSessionAdapter, PiAgentSessionSnapshot } from '../pi-chat/PiAgentSessionAdapter'
 import { ErrorCode } from '../../shared/error-codes'
 import { decideInputAssetIntake } from '../inputAssetIntake'
+import type { PiChatEvent, PromptPayload } from '../../shared/chat'
 
 const CTX: SessionCtx = { workspaceId: 'workspace-test', userId: 'user-test' }
 
 describe('createAgent', () => {
   it('exposes exactly the nine-member facade without Fastify', () => {
-    const agent = createAgent({
-      runtime: 'none',
-      harnessFactory: createFakeHarnessFactory(),
+    const agent = createCoreAgent({
+      runtimeFactory: createCoreRuntimeFactory(),
     })
 
     expect(Object.keys(agent).sort()).toEqual([
@@ -37,6 +39,17 @@ describe('createAgent', () => {
       'stop',
       'stream',
     ])
+  })
+
+  it('core public runtimeFactory path runs without the server wrapper', async () => {
+    const agent = createCoreAgent({
+      runtimeFactory: createCoreRuntimeFactory({ autoCompletePrompt: true }),
+    })
+
+    const events = await collectEvents(agent.send({ content: 'hello from core', ctx: CTX }))
+
+    expect(events.map((event) => event.chunk.type)).toEqual(['agent-start', 'agent-end'])
+    await agent.dispose()
   })
 
   it('rejects sessions override with the default harness to avoid split persistence', () => {
@@ -630,6 +643,20 @@ describe('createAgent', () => {
   })
 })
 
+function createCoreRuntimeFactory(options: { autoCompletePrompt?: boolean } = {}): AgentCoreRuntimeFactory {
+  const sessions = new MemorySessionStore()
+  const service = new InjectedCorePiChatService(sessions, options)
+  return async () => ({
+    harness: {
+      id: 'core-injected',
+      placement: 'server',
+      sessions,
+    },
+    sessionStore: sessions,
+    service,
+  })
+}
+
 function createFakeHarnessFactory(options: { autoCompletePrompt?: boolean; eventsPerPrompt?: number; seedSessions?: string[] } = {}) {
   const sessions = new MemorySessionStore()
   for (const sessionId of options.seedSessions ?? []) sessions.seed(sessionId, CTX)
@@ -777,8 +804,100 @@ class ScopedSessionStore extends MemorySessionStore {
   }
 }
 
+class InjectedCorePiChatService implements PiChatSessionService {
+  private readonly subscribers = new Map<string, Set<PiChatEventSubscriber>>()
+  private readonly latestSeq = new Map<string, number>()
+  private turns = 0
+
+  constructor(
+    private readonly sessions: MemorySessionStore,
+    private readonly options: { autoCompletePrompt?: boolean },
+  ) {}
+
+  async createSession(ctx: PiSessionRequestContext, init?: { title?: string }) {
+    return this.sessions.create(toTestSessionCtx(ctx), init)
+  }
+
+  async deleteSession(ctx: PiSessionRequestContext, sessionId: string): Promise<void> {
+    await this.sessions.delete(toTestSessionCtx(ctx), sessionId)
+    this.subscribers.delete(sessionId)
+    this.latestSeq.delete(sessionId)
+  }
+
+  async readState(_ctx: PiSessionRequestContext, sessionId: string) {
+    return {
+      protocolVersion: 1 as const,
+      sessionId,
+      seq: this.latestSeq.get(sessionId) ?? 0,
+      status: 'idle' as const,
+      messages: [],
+      queue: { followUps: [] },
+      followUpMode: 'one-at-a-time' as const,
+    }
+  }
+
+  async subscribe(ctx: PiSessionRequestContext, sessionId: string, _cursor: number, subscriber: PiChatEventSubscriber) {
+    await this.sessions.load(toTestSessionCtx(ctx), sessionId)
+    const subscribers = this.subscribers.get(sessionId) ?? new Set<PiChatEventSubscriber>()
+    subscribers.add(subscriber)
+    this.subscribers.set(sessionId, subscribers)
+    return {
+      type: 'ok' as const,
+      unsubscribe: () => subscribers.delete(subscriber),
+    }
+  }
+
+  async prompt(_ctx: PiSessionRequestContext, sessionId: string, payload: PromptPayload) {
+    const turnId = `core-turn-${++this.turns}`
+    this.publish(sessionId, { type: 'agent-start', seq: this.nextSeq(sessionId), turnId })
+    if (this.options.autoCompletePrompt === true) {
+      this.publish(sessionId, { type: 'agent-end', seq: this.nextSeq(sessionId), turnId, status: 'ok' })
+    }
+    return { accepted: true as const, cursor: this.latestSeq.get(sessionId) ?? 0, clientNonce: payload.clientNonce }
+  }
+
+  async followUp(_ctx: PiSessionRequestContext, sessionId: string, payload: { clientNonce: string; clientSeq: number }) {
+    return {
+      accepted: true as const,
+      cursor: this.latestSeq.get(sessionId) ?? 0,
+      clientNonce: payload.clientNonce,
+      clientSeq: payload.clientSeq,
+      queued: true as const,
+    }
+  }
+
+  async clearQueue(_ctx: PiSessionRequestContext, sessionId: string) {
+    return { accepted: true as const, cursor: this.latestSeq.get(sessionId) ?? 0, cleared: 0 }
+  }
+
+  async interrupt(_ctx: PiSessionRequestContext, sessionId: string) {
+    return { accepted: true as const, cursor: this.latestSeq.get(sessionId) ?? 0 }
+  }
+
+  async stop(_ctx: PiSessionRequestContext, sessionId: string) {
+    return { accepted: true as const, cursor: this.latestSeq.get(sessionId) ?? 0, stopped: true as const, clearedQueue: [] }
+  }
+
+  private nextSeq(sessionId: string): number {
+    const next = (this.latestSeq.get(sessionId) ?? 0) + 1
+    this.latestSeq.set(sessionId, next)
+    return next
+  }
+
+  private publish(sessionId: string, event: PiChatEvent): void {
+    for (const subscriber of this.subscribers.get(sessionId) ?? []) subscriber(event)
+  }
+}
+
 function normalizeTestCtx(ctx: SessionCtx | undefined): SessionCtx | undefined {
   return !ctx?.workspaceId && !ctx?.userId ? undefined : { workspaceId: ctx.workspaceId, userId: ctx.userId }
+}
+
+function toTestSessionCtx(ctx: PiSessionRequestContext): SessionCtx {
+  return {
+    workspaceId: ctx.workspaceId,
+    userId: ctx.authSubject,
+  }
 }
 
 function sameTestCtx(a: SessionCtx | undefined, b: SessionCtx | undefined): boolean {
