+++ packages/agent/src/server/createAgentApp.ts
@@ -14,7 +14,7 @@ import { loadPlugins } from './harness/pi-coding-agent/pluginLoader'
 import { buildFilesystemAgentTools } from './tools/filesystem'
 import { buildHarnessAgentTools } from './tools/harness'
 import { createAuthMiddleware } from './http/middleware'
-import type { PiChatSessionService } from './http/routes/piChat'
+import type { PiChatSessionService } from '../core/piChatSessionService'
 import { InMemorySessionChangesTracker } from './http/sessionChangesTracker'
 import { createRuntimeReadyStatusTracker } from './runtime/modeReadiness'
 import { ReadyStatusTracker } from './runtime/readyStatus'
