+++ packages/agent/src/server/registerAgentRoutes.ts
@@ -26,7 +26,7 @@ import type { ToolReadinessState } from './catalog/toolReadiness'
 import { buildFilesystemAgentTools } from './tools/filesystem'
 import { buildHarnessAgentTools } from './tools/harness'
 import { buildUploadAgentTools } from './tools/upload'
-import type { PiChatSessionService } from './http/routes/piChat'
+import type { PiChatSessionService } from '../core/piChatSessionService'
 import type { ReloadHookResult } from './http/routes/reload'
 import { InMemorySessionChangesTracker } from './http/sessionChangesTracker'
 import { ReadyStatusTracker } from './runtime/readyStatus'
